requirements.txt
opencv-python==4.12.0.88
numpy==2.2.6
Pillow==11.1.0
matplotlib==3.10.3
